requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via -r requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via uvicorn
coloredlogs==15.0.1
    # via onnxruntime
databases==0.9.0
    # via -r requirements.in
dnspython==2.7.0
    # via pymongo
fastapi==0.115.6
    # via -r requirements.in
flatbuffers==24.12.23
    # via onnxruntime
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
mpmath==1.3.0
    # via sympy
numpy==1.26.4
    # via
    #   -r requirements.in
    #   onnxruntime
onnxruntime==1.20.1
    # via -r requirements.in
packaging==24.2
    # via
    #   onnxruntime
    #   pytest
pluggy==1.5.0
    # via pytest
protobuf==5.29.2
    # via onnxruntime
psycopg2==2.9.10
    # via -r requirements.in
pydantic==2.10.4
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
pymongo==4.10.1
    # via -r requirements.in
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.25.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   databases
starlette==0.41.3
    # via fastapi
sympy==1.13.3
    # via onnxruntime
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
    # via -r requirements.in
